+++ services/core/methods/api_v1/controllers/accounts.js
@@ -20,7 +20,7 @@ const { StatusCodes: { NOT_FOUND } } = HTTP;
 const { isEmptyArray } = Utils.Data;
 
 const CoreService = require('../../../shared/core');
-const { getAccountKnowledge } = require('../../../shared/core/knownAccounts');
+const { getAccountKnowledge } = require('../../../shared/knownAccounts');
 const { parseToJSONCompatObj } = require('../../../shared/jsonTools');
 
 const logger = Logger();
